Limit the number of threads PyArrow is allowed to use for conversion of a Feather database to a pandas dataframe.
Limit the number of threads PyArrow is allowed to use for conversion
of a Feather database to a pandas dataframe to 4 threads. Using more
threads usually does not make loading a Feather database much faster.
On big machines it even slows down the loading due to too much
concurency and memory access.

To override the default number of 4 threads (e.g. 8 threads), the
following environment variable can be set before starting the python
interpreter.

    PYARROW_THREADS=8

The number of threads can also be set by executing the following
python code:

    PyArrowThreads.set_nbr_pyarrow_threads(8)